LPS-162959 Move Tooltip inside SpacingBox
