+++ modules/apps/layout/layout-content-page-editor-web/src/main/resources/META-INF/resources/page_editor/common/components/SpacingBox.js
@@ -14,6 +14,8 @@
 
 import ClayButton from '@clayui/button';
 import ClayDropDown from '@clayui/drop-down';
+import ClayTooltip from '@clayui/tooltip';
+import {ReactPortal} from '@liferay/frontend-js-react-web';
 import classNames from 'classnames';
 import React, {useEffect, useRef, useState} from 'react';
 
@@ -22,7 +24,6 @@ import isValidStyleValue from '../../app/utils/isValidStyleValue';
 import {LengthField} from '../../common/components/LengthField';
 import {useId} from '../../core/hooks/useId';
 import {useStyleBook} from '../../plugins/page-design-options/hooks/useStyleBook';
-import {Tooltip} from './Tooltip';
 
 /**
  * These elements must be sorted from the most outer circle to the most inner
@@ -54,6 +55,8 @@ const REVERSED_POSITION = {
 
 const BUTTON_CLASSNAME = 'page-editor__spacing-selector__button';
 const DROPDOWN_CLASSNAME = 'page-editor__spacing-selector__dropdown';
+
+const TOOLTIP_SHOW_DELAY = 600;
 
 export default function SpacingBox({
 	canSetCustomValue,
@@ -346,6 +349,59 @@ function SpacingOptionValue({
 	]);
 
 	return value === undefined ? '' : value;
+}
+
+export function Tooltip({hoverElement, id: tooltipId, label, positionElement}) {
+	const [tooltipStyle, setTooltipStyle] = useState(null);
+
+	useEffect(() => {
+		if (!hoverElement || !positionElement) {
+			return;
+		}
+
+		let showTimeoutId;
+
+		const handleMouseLeave = () => {
+			clearTimeout(showTimeoutId);
+			setTooltipStyle(null);
+		};
+
+		const handleMouseOver = () => {
+			clearTimeout(showTimeoutId);
+
+			showTimeoutId = setTimeout(() => {
+				const rect = positionElement.getBoundingClientRect();
+
+				setTooltipStyle({
+					left: rect.left + rect.width / 2,
+					top: rect.top,
+				});
+			}, TOOLTIP_SHOW_DELAY);
+		};
+
+		hoverElement.addEventListener('mouseleave', handleMouseLeave);
+		hoverElement.addEventListener('mouseover', handleMouseOver);
+
+		return () => {
+			clearTimeout(showTimeoutId);
+			hoverElement.removeEventListener('mouseleave', handleMouseLeave);
+			hoverElement.removeEventListener('mouseover', handleMouseOver);
+		};
+	}, [hoverElement, positionElement]);
+
+	return tooltipStyle ? (
+		<ReactPortal className="cadmin">
+			<ClayTooltip
+				alignPosition="top"
+				className="page-editor__tooltip position-fixed"
+				id={tooltipId}
+				show
+				style={tooltipStyle}
+			>
+				{label}
+			</ClayTooltip>
+		</ReactPortal>
+	) : null;
 }
 
 function capitalize(str) {
